Add extra slashes before path names to support paths with dates
Some of our dataset are stored in paths with dates having the format 12.10.2024 and these no longer process correctly going to the lastest version of the n5 libraries. By adding two extra // to all paths, this issues seems to be resolved. Not a great solution, but appears the only option at the moment.